got home activity to draw from the lamp database to display app data
hey so i made it so that home activity draws from the database and sends the necessary information to the Donate page when you click on a user in the recycler view.
I was learning all of this and had to restructure some of the classes and files and such so there is probably some redundant or trivial files added here that can be deleted. next task is fixing the DonateActivity so that it shows the correct data.