Vibe coded the search functionality
got the search functionality to run. I also changed the edittext into  a searchview which has messed up the colours I guess. I don't really get it but what matters is that it is done. now it is time to get to the donate page and I think I want to reformat the code for readability and understanding